+++ Model/Ui/GenericGiftcardConfigProvider.php
@@ -21,6 +21,10 @@ use Magento\Checkout\Model\ConfigProviderInterface;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\View\Asset\Repository as AssetRepository;
 
+/**
+ * @SuppressWarnings(PHPMD.NumberOfChildren)
+ * @SuppressWarnings(PHPMD.CookieAndSessionMisuse)
+ */
 class GenericGiftcardConfigProvider extends GenericConfigProvider
 {
     /**