+++ Model/Vault.php
@@ -42,6 +42,9 @@ use MultiSafepay\ConnectCore\Util\JsonHandler;
 use MultiSafepay\ConnectCore\Util\VaultUtil;
 use \MultiSafepay\ConnectCore\Api\PaymentTokenInterface as MultiSafepayPaymentTokenInterface;
 
+/**
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ */
 class Vault
 {
     public const VAULT_GATEWAYS = [