+++ Test/Integration/Model/Api/Builder/OrderRequestBuilder/GatewayInfoBuilderTest.php
@@ -33,6 +33,9 @@ use MultiSafepay\ConnectCore\Model\Ui\Gateway\IdealConfigProvider;
 use MultiSafepay\ConnectCore\Model\Ui\Gateway\VisaConfigProvider;
 use MultiSafepay\ConnectCore\Test\Integration\Payment\AbstractTransactionTestCase;
 
+/**
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ */
 class GatewayInfoBuilderTest extends AbstractTransactionTestCase
 {
     /**